It seems like this approach is maybe slightly less performant than the one in [master@15d7686](https://github.com/mapbox/fuzzy-phrase/tree/15d7686ea903345d337dbec59639242096e51e58) and the one in #25?

| benchmark                    | covers                                         | master   | PR #25   | this PR  |
|------------------------------|------------------------------------------------|----------|----------|----------|
| phrase/exact_contains        | `PhraseSet::contains`                          | 2.3250us | 1.7847us | 2.5476us |
| phrase/exact_contains_prefix | `PhraseSet::contains_prefix` (all full words)  | 2.2820us | 2.4375us | 2.8590us |
| phrase/range_contains_prefix | `PhraseSet::contains_prefix` (terminal prefix) | 2.5716us | 2.4619us | 3.1849us |