@boblannon 2018-05-30
tried this again on an ECS, hoping to avoid interference from my laptop's processor magic. master is 15d7686.

| bench                        | master                     | #25             | #26            |
|------------------------------|----------------------------|-----------------|----------------|
| phrase/exact_contains        | [1.7112 **1.7368** 1.7603] | 1.6591 (-7.0%) | 1.7758 (+2.2%) |
| phrase/exact_contains_prefix | [1.5976 **1.6251** 1.6526] | 1.6429 (-2.5%)  | 1.7191 (+5.8%) |
| phrase/range_contains_prefix | [1.7661 **1.7966** 1.8218] | 1.7718 (-1.8%)  | 1.8766 (+4.5%) |
| phrase/range_fst_range       | [2.7078 **2.7440** 2.7753] | 2.8037 (+2.0%)  | 2.9959 (+9.1%) |

Note: if you look at the details of experiment2's bench run below, you'll notice that the change percentages differ from what's above.  That's because I ran four benchmark runs: 

1. bench on master
2. bench on #25 
3. bench on master
4. bench on #26 

In other words, I re-ran on master in between so that its results would become the `base` set and I'd get appropriate comparisons from criterion's output.  Apparently the speeds differed in the two runs against master.  Unfortunately the range between those two runs is 0.10us, which makes me wonder if any of these results is really significant...


<details>

<summary>`phrase-prefix-experiment` details</summary>


baseline (master):

```
phrase/exact_contains   time:   [1.7112 us 1.7368 us 1.7603 us]
phrase/exact_contains_prefix
                        time:   [1.5976 us 1.6251 us 1.6526 us]
Found 12 outliers among 100 measurements (12.00%)
  12 (12.00%) low mild
phrase/range_contains_prefix
                        time:   [1.7661 us 1.7966 us 1.8218 us]
phrase/range_fst_range  time:   [2.7078 us 2.7440 us 2.7753 us]
```

comparison (#25):

```


phrase/exact_contains   time:   [1.6181 us 1.6591 us 1.6939 us]
                        change: [-10.556% -6.9998% -3.5479%] (p = 0.00 < 0.05)
                        Performance has improved.
phrase/exact_contains_prefix
                        time:   [1.6241 us 1.6429 us 1.6590 us]
                        change: [-5.6044% -2.5239% +0.5874%] (p = 0.11 > 0.05)
                        No change in performance detected.
phrase/range_contains_prefix
                        time:   [1.7363 us 1.7718 us 1.8018 us]
                        change: [-5.6779% -1.8348% +2.0410%] (p = 0.36 > 0.05)
                        No change in performance detected.
phrase/range_fst_range  time:   [2.7780 us 2.8037 us 2.8266 us]
                        change: [-0.5300% +1.9998% +4.5943%] (p = 0.13 > 0.05)
                        No change in performance detected.
```

</details>

<details>

<summary>`phrase-prefix-experiment2` details</summary>


baseline (master). change stats are against #25, so ignore them:

```
phrase/exact_contains   time:   [1.6217 us 1.6472 us 1.6678 us]
                        change: [-3.0151% +0.9240% +4.8513%] (p = 0.65 > 0.05)
                        No change in performance detected.
phrase/exact_contains_prefix
                        time:   [1.6323 us 1.6581 us 1.6841 us]
                        change: [+4.0354% +7.2603% +10.646%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 12 outliers among 100 measurements (12.00%)
  8 (8.00%) low mild
  4 (4.00%) high mild
phrase/range_contains_prefix
                        time:   [1.8538 us 1.8876 us 1.9217 us]
                        change: [+4.4587% +8.5075% +12.696%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
phrase/range_fst_range  time:   [2.7508 us 2.7831 us 2.8106 us]
                        change: [-3.7224% -1.2192% +1.3282%] (p = 0.35 > 0.05)
                        No change in performance detected.
```

comparison (#26):

```
phrase/exact_contains   time:   [1.7572 us 1.7758 us 1.7931 us]
                        change: [+7.9948% +11.406% +14.950%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) low mild
phrase/exact_contains_prefix
                        time:   [1.7048 us 1.7191 us 1.7354 us]
                        change: [-1.9942% +0.9579% +3.8251%] (p = 0.52 > 0.05)
                        No change in performance detected.
Found 17 outliers among 100 measurements (17.00%)
  13 (13.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
phrase/range_contains_prefix
                        time:   [1.8539 us 1.8766 us 1.8960 us]
                        change: [-5.1858% -1.6473% +2.0615%] (p = 0.39 > 0.05)
                        No change in performance detected.
phrase/range_fst_range  time:   [2.9857 us 2.9959 us 3.0075 us]
                        change: [+11.239% +13.592% +16.031%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 15 outliers among 100 measurements (15.00%)
  8 (8.00%) low severe
  1 (1.00%) low mild
  6 (6.00%) high mild
```

</details>