@apendleton 2018-06-07
Closing in favor of #25